+++ src/components/ComponentRuntimeException.php
@@ -1,0 +1,9 @@
+<?php
+namespace VovanVE\HtmlTemplate\components;
+
+/**
+ * @since 0.3.2
+ */
+class ComponentRuntimeException extends ComponentException
+{
+}