+++ tests/unit/compile/CompilerTest.php
@@ -9,6 +9,7 @@ use VovanVE\HtmlTemplate\source\memory\TemplateString;
 use VovanVE\HtmlTemplate\tests\helpers\BaseTestCase;
 use VovanVE\HtmlTemplate\tests\helpers\conversion\Expect;
 use VovanVE\HtmlTemplate\tests\helpers\CounterStepperComponent;
+use VovanVE\HtmlTemplate\tests\helpers\FailureComponent;
 use VovanVE\HtmlTemplate\tests\helpers\RuntimeCounter;
 use VovanVE\HtmlTemplate\tests\helpers\StringConversionTestTrait;
 use VovanVE\HtmlTemplate\tests\helpers\TestComponent;
@@ -42,6 +43,7 @@ class CompilerTest extends BaseTestCase
         $cache = new CacheStrings(__FUNCTION__ . '_%{hash}', __CLASS__);
         $runtime = (new RuntimeCounter)
             ->addComponents([
+                'Failure' => FailureComponent::class,
                 'TestComponent' => TestComponent::class,
                 'Step' => new CounterStepperComponent(),
             ]);